backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
firebase-admin==6.2.0
google-cloud-firestore==2.15.0
google-cloud-storage==2.10.0
google-cloud-aiplatform>=1.60.0
vertexai>=1.60.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-translate==3.12.1
google-auth==2.23.4
google-cloud-speech==2.21.0
pillow==10.1.0
aiofiles==23.2.1
